+++ web/app/(app)/dashboard/(components)/alerts/community-alert.tsx
