+++ web/app/(app)/dashboard/(components)/devices/device-list.tsx
