+++ web/app/(app)/dashboard/(components)/devices/update-app-helpers.ts
