+++ web/app/(app)/dashboard/(components)/search/search-registry.ts
