+++ web/app/(app)/dashboard/(components)/search/search-trigger.tsx
