+++ web/app/(app)/dashboard/account/billing/page.tsx
@@ -1,4 +1,4 @@
-import SubscriptionInfo from '../../(components)/subscription-info'
+import SubscriptionInfo from '../../(components)/billing/subscription-info'
 
 // Billing & plan: current subscription, usage limits and plan CTAs.
 export default function BillingPage() {
