+++ web/app/(app)/dashboard/layout.tsx
@@ -3,17 +3,17 @@
 import { useState } from 'react'
 import Link from 'next/link'
 import { usePathname } from 'next/navigation'
-import AccountDeletionAlert from './(components)/account-deletion-alert'
-import UpgradeToProAlert from './(components)/upgrade-to-pro-alert'
-import UpdateAppModal from './(components)/update-app-modal'
-import UpdateAppNotificationBar from './(components)/update-app-notification-bar'
-import VerifyEmailAlert from './(components)/verify-email-alert'
-import PastDueBillingAlert from './(components)/past-due-billing-alert'
+import AccountDeletionAlert from './(components)/alerts/account-deletion-alert'
+import UpgradeToProAlert from './(components)/alerts/upgrade-to-pro-alert'
+import UpdateAppModal from './(components)/devices/update-app-modal'
+import UpdateAppNotificationBar from './(components)/devices/update-app-notification-bar'
+import VerifyEmailAlert from './(components)/alerts/verify-email-alert'
+import PastDueBillingAlert from './(components)/alerts/past-due-billing-alert'
 import { SurveyModal } from '@/components/shared/survey-modal'
 import Footer from '@/components/shared/footer'
 import ThemeToggle from '@/components/shared/theme-toggle'
-import CommandMenu from './(components)/command-menu'
-import SearchTrigger from './(components)/search-trigger'
+import CommandMenu from './(components)/search/command-menu'
+import SearchTrigger from './(components)/search/search-trigger'
 import {
   isNavItemActive,
   mobileNavItems,
